Update traveling_salesman.c
Updated the code so that the tours are being save. Able to get the best tour and print only those out at the end of the program. Moved some of the printing and simple functions the could be refactored out into their own functions to make things more reusable. Seems to be working pretty well over all. The current structure should be easier to reimplement in the other parallel versions that we need to make up.